+++ jobmon_gui/src/vite-env.d.ts
@@ -1,2 +1,7 @@
 /// <reference types="vite/client" />
 declare const APP_VERSION: string;
+
+declare module 'plotly.js-cartesian-dist' {
+    import Plotly from 'plotly.js';
+    export default Plotly;
+}